Start refactoring grapheditor class into smaller renderers
